+++ CHANGELOG.md
@@ -1,40 +1,78 @@
+## Fix [v1.4.7] 2023-06-29
+
+- change status color plot card.
+
+## Fix [v1.4.6] 2023-06-22
+
+- added modal profile status rejected & inactive
+
+## Fix [v1.4.5] 2023-06-21
+
 ## Fix [v1.4.4] 2023-06-12
+
 - fix calculate price with coupon
+
 ## Fix [v1.4.3] 2023-06-07
+
 - fix plotName & lat long
+
 ## Fix [v1.4.2] 2023-06-06
+
 - fix coupon detail
+
 ## Fix [v1.4.1] 2023-06-06
+
 - fix task img droner used in main screen
+
 ## Feature [v1.4.0] 2023-05-16
+
 - add history point feature
+
 ## Fix [v1.3.4] 2023-04-27
+
 - fix app error
 - fix offline code
+
 ## Fix [v1.3.3] 2023-04-27
+
 - fix price display
 - fix bug selet prepare check box in select target screen
+
 ## Fix [v1.3.2] 2023-04-26
+
 - comment review
+
 ## Fix [v1.3.1] 2023-04-24
+
 - fixed fetching data guru kaset
+
 ## Feature [v1.3.0] 2023-04-24
-- added guru kaset feature 
+
+- added guru kaset feature
 
 ## Feature [v1.2.5] 2023-04-05
+
 - add promotion feature for droner service
+
 ## Fix [v1.2.4] 2023-04-05
-- fixed time onboarding 
+
+- fixed time onboarding
+
 ## Fix [v1.2.3] 2023-04-03
+
 - fixed bug fav in main screen
 - added flow favorite in detail droner screen
 - added navigate to detail droner screen
+
 ## Fix [v1.2.2] 2023-03-30
+
 - fixed favorite droner
 - fixed loading droner detail review
 
 ## Feature [v1.2.1] 2023-03-28
+
 - add head droner card select booking
 
 ## Feature [v1.2.0] 2023-03-27
+
 - feature select booking